+++ app/src/main/java/ru/yellowshark/dsr_weather/domain/repository/Repository.kt
@@ -27,7 +27,7 @@ interface Repository {
 
     fun setLocationName(name: String)
 
-    fun setCoordinates(lon: Long, lat: Long)
+    fun setCoordinates(lon: Double, lat: Double)
 
     fun getUnit(): String
 }
